Delete unused facade, classes and rename LaravelUssdServiceProvider.php to UssdServiceProvider.php
