--- src/LaravelUssd.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace CyberSai\LaravelUssd;
-
-class LaravelUssd
-{
-    // Build wonderful things
-}
